Refactor: searchbar: move state files and test to redux folder
