+++ src/components/SearchBar/redux/__tests__/searchSlice.test.ts
@@ -1,4 +1,4 @@
-import searchReducer, { searchActions } from 'components/SearchBar/searchSlice';
+import searchReducer, { searchActions } from 'components/SearchBar/redux/searchSlice';
 
 describe('searchReducer', () => {
     it('should return state with set value', () => {
