+++ src/components/SearchBar/redux/searchBarSelectors.ts
