+++ src/components/SearchBar/redux/searchSlice.ts
